pyproject.toml
[project]
name = "autobi-powerbi-mcp"
version = "0.1.0"
description = "Cross-project MCP server for Power BI and Microsoft Fabric operations"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0",
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "azure-identity>=1.15",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-mock>=3.12",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
    "pre-commit>=4.0.0",
    "pip-tools>=7.4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=src --cov-report=term-missing --cov-fail-under=80"
markers = [
    "integration: tests requiring live Fabric API access",
    "slow: tests that take more than 5 seconds",
]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
include = ["src"]
exclude = ["scripts", "tests", "**/__pycache__"]
